Gui: Reorganize the navigation files into a top `Navigation` folder.
