+++ src/Gui/Navigation/GestureNavigationStyle-state-machine-diagram.docx
