+++ src/Gui/Navigation/GestureNavigationStyle.cpp
