+++ src/Gui/Navigation/NavigationAnimation.cpp
