+++ src/Gui/Navigation/NavigationAnimator.cpp
