+++ src/Gui/Navigation/NavigationStyle.cpp
@@ -42,14 +42,14 @@
 #include <Base/Interpreter.h>
 #include <App/Application.h>
 
-#include "NavigationStyle.h"
-#include "NavigationStylePy.h"
+#include "Navigation/NavigationStyle.h"
+#include "Navigation/NavigationStylePy.h"
 #include "Application.h"
 #include "Inventor/SoMouseWheelEvent.h"
 #include "MenuManager.h"
 #include "MouseSelection.h"
-#include "NavigationAnimator.h"
-#include "NavigationAnimation.h"
+#include "Navigation/NavigationAnimator.h"
+#include "Navigation/NavigationAnimation.h"
 #include "View3DInventorViewer.h"
 
 using namespace Gui;
