+++ src/Gui/Navigation/NavigationStyle.h
